Add unit test project

+++ Practeece.GolfHoles/models/GolfHole.cs
@@ -8,17 +8,17 @@ namespace Practeece.GolfHoles.models
 {
     public class GolfHole
     {
-        public int HoleId { get; set; }
+        public byte HoleId { get; set; }
 
         public int Distance { get; set; }
 
-        public int Par { get; set; }
+        public byte Par { get; set; }
 
         public string Skill { get; set; }
 
         public string Hazard { get; set; }
 
-        public int AverageSwingsToGreen { get; set; }
+        public byte AverageSwingsToGreen { get; set; }
 
     }
 }
