+++ Practeece.GolfHoles/models/GolfSwing.cs
@@ -10,7 +10,7 @@ namespace Practeece.GolfHoles.models
     {
         public GolfHole GolfHole { get; set; }
 
-        public int SwingCount { get; set; }
+        public byte SwingCount { get; set; }
 
         public string SkillLevel { get; set; }
     }
